requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
captcha==0.7.1
cffi==1.17.1
colorama==0.4.6
colorlog==6.9.0
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
multidict==6.5.0
pillow==11.2.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyNaCl==1.5.0
python-dotenv==1.1.1
pytube==15.0.0
pytubefix==9.4.1
topggpy==2.0.0a0
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
